## Description

<!-- Describe the goal and purpose of this PR. -->

Enable the processors/appliers to work with ID on group deletion

## Related issues

closes #30025

+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/identity/GroupDeleteProcessor.java
@@ -32,7 +32,7 @@ import io.camunda.zeebe.stream.api.state.KeyGenerator;
 public class GroupDeleteProcessor implements DistributedTypedRecordProcessor<GroupRecord> {
 
   private static final String GROUP_NOT_FOUND_ERROR_MESSAGE =
-      "Expected to delete group with key '%s', but a group with this key does not exist.";
+      "Expected to delete group with ID '%s', but a group with this ID does not exist.";
   private final GroupState groupState;
   private final AuthorizationState authorizationState;
   private final AuthorizationCheckBehavior authCheckBehavior;
@@ -61,18 +61,10 @@ public class GroupDeleteProcessor implements DistributedTypedRecordProcessor<Gro
   @Override
   public void processNewCommand(final TypedRecord<GroupRecord> command) {
     final var record = command.getValue();
-    final var groupKey = record.getGroupKey();
-    final var persistedRecord = groupState.get(groupKey);
-    if (persistedRecord.isEmpty()) {
-      final var errorMessage = GROUP_NOT_FOUND_ERROR_MESSAGE.formatted(groupKey);
-      rejectionWriter.appendRejection(command, RejectionType.NOT_FOUND, errorMessage);
-      responseWriter.writeRejectionOnCommand(command, RejectionType.NOT_FOUND, errorMessage);
-      return;
-    }
-
+    final var groupId = record.getGroupId();
     final var authorizationRequest =
         new AuthorizationRequest(command, AuthorizationResourceType.GROUP, PermissionType.DELETE)
-            .addResourceId(persistedRecord.get().getName());
+            .addResourceId(groupId);
     final var isAuthorized = authCheckBehavior.isAuthorized(authorizationRequest);
     if (isAuthorized.isLeft()) {
       final var rejection = isAuthorized.getLeft();
@@ -80,7 +72,17 @@ public class GroupDeleteProcessor implements DistributedTypedRecordProcessor<Gro
       responseWriter.writeRejectionOnCommand(command, rejection.type(), rejection.reason());
       return;
     }
-    record.setName(persistedRecord.get().getName());
+
+    final var persistedRecord = groupState.get(groupId);
+    if (persistedRecord.isEmpty()) {
+      final var errorMessage = GROUP_NOT_FOUND_ERROR_MESSAGE.formatted(groupId);
+      rejectionWriter.appendRejection(command, RejectionType.NOT_FOUND, errorMessage);
+      responseWriter.writeRejectionOnCommand(command, RejectionType.NOT_FOUND, errorMessage);
+      return;
+    }
+
+    final var persistedGroup = persistedRecord.get();
+    final var groupKey = persistedGroup.getGroupKey();
 
     removeAssignedEntities(record);
     deleteAuthorizations(record);
@@ -99,7 +101,7 @@ public class GroupDeleteProcessor implements DistributedTypedRecordProcessor<Gro
   public void processDistributedCommand(final TypedRecord<GroupRecord> command) {
     final var record = command.getValue();
     groupState
-        .get(record.getGroupKey())
+        .get(record.getGroupId())
         .ifPresentOrElse(
             group -> {
               removeAssignedEntities(command.getValue());
@@ -135,10 +137,9 @@ public class GroupDeleteProcessor implements DistributedTypedRecordProcessor<Gro
   }
 
   private void deleteAuthorizations(final GroupRecord record) {
-    final var groupKey = record.getGroupKey();
+    final var groupId = record.getGroupId();
     final var authorizationKeysForGroup =
-        authorizationState.getAuthorizationKeysForOwner(
-            AuthorizationOwnerType.GROUP, String.valueOf(groupKey));
+        authorizationState.getAuthorizationKeysForOwner(AuthorizationOwnerType.GROUP, groupId);
 
     authorizationKeysForGroup.forEach(
         authorizationKey -> {
